fix(do.go): indexing bug in do.go
Fixed a bug in do.go and list.go. When items were completed and removed
from the DB, they would be printed with their DB ID, which doesn't
begin at '1'. Added logic in list.go to print from 1 as well as logic
in do.go to retrieve the appropriate task ID for the DB so the user
could supply the list item for appropriate task completion